+++ make/hotspot/lib/JvmFeatures.gmk
@@ -118,6 +118,7 @@ ifneq ($(call check-jvm-feature, cds), true)
   JVM_EXCLUDE_FILES += \
       archiveBuilder.cpp \
       archiveUtils.cpp \
+      cdsoffsets.cpp \
       classListParser.cpp \
       classLoaderDataShared.cpp \
       classLoaderExt.cpp \
