+++ src/hotspot/share/cds/classListParser.cpp
@@ -26,11 +26,11 @@
 #include "jvm.h"
 #include "jimage.hpp"
 #include "cds/archiveUtils.hpp"
+#include "cds/classListParser.hpp"
+#include "cds/lambdaFormInvokers.hpp"
 #include "cds/metaspaceShared.hpp"
-#include "classfile/classListParser.hpp"
 #include "classfile/classLoaderExt.hpp"
 #include "classfile/javaClasses.inline.hpp"
-#include "classfile/lambdaFormInvokers.hpp"
 #include "classfile/symbolTable.hpp"
 #include "classfile/systemDictionary.hpp"
 #include "classfile/systemDictionaryShared.hpp"
