+++ src/hotspot/share/cds/classListWriter.hpp
