+++ src/hotspot/share/cds/compactHashtable.cpp
@@ -25,8 +25,8 @@
 #include "precompiled.hpp"
 #include "jvm.h"
 #include "cds/archiveBuilder.hpp"
+#include "cds/compactHashtable.hpp"
 #include "cds/heapShared.inline.hpp"
-#include "classfile/compactHashtable.hpp"
 #include "classfile/javaClasses.hpp"
 #include "logging/logMessage.hpp"
 #include "memory/metadataFactory.hpp"
