+++ src/hotspot/share/cds/lambdaFormInvokers.hpp
