+++ src/hotspot/share/classfile/stringTable.cpp
@@ -23,8 +23,10 @@
  */
 
 #include "precompiled.hpp"
+#include "cds/archiveBuilder.hpp"
+#include "cds/compactHashtable.hpp"
+#include "cds/heapShared.inline.hpp"
 #include "classfile/altHashing.hpp"
-#include "classfile/compactHashtable.hpp"
 #include "classfile/javaClasses.inline.hpp"
 #include "classfile/stringTable.hpp"
 #include "classfile/vmClasses.hpp"
@@ -34,8 +36,6 @@
 #include "logging/log.hpp"
 #include "logging/logStream.hpp"
 #include "memory/allocation.inline.hpp"
-#include "cds/archiveBuilder.hpp"
-#include "cds/heapShared.inline.hpp"
 #include "memory/resourceArea.hpp"
 #include "memory/universe.hpp"
 #include "oops/access.inline.hpp"
